chore(93): restructure component files
