+++ hivemq-edge/src/frontend/src/modules/ProtocolAdapters/components/templates/ObjectFieldTemplate.tsx
